collabtree-backend/requirements.txt
fastapi>=0.68.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.9,<3.0.0
pydantic[email]>=1.8.2
alembic>=1.7.1
python-dotenv>=0.19.0
bcrypt>=4.0.1